Added support for AVX2 (256-bit) and AVX512 (512-bit) operations into pipeline generator
The support is currently enabled only for few fetchers (solid fills, blits, and linear
gradient). It's planned to extend it in the future to cover all possible fetch operations.

+++ src/blend2d/opentype/otglyf.cpp
@@ -332,7 +332,7 @@ static BLResult BL_CDECL getGlyphOutlines(
               if (BL_UNLIKELY(n > ttVertexCount - i))
                 goto InvalidData;
 
-              xyCoordinatesSize += n * vertexSize;
+              xyCoordinatesSize += uint32_t(n) * vertexSize;
               offCurveSplineCount += n * size_t((f & Simple::kOnCurvePoint) == 0);
 
               BLMemOps::fillSmall(fDataPtr + i, uint8_t(f), n);
